Added support to notify events through a notifier
Now we can import the notification_center from guiapi and notify any UI
about certain events. Most of them are triggered by the model controller
and the workspace, but it can be used anywhere.

Every action has the responsabilty of trigger the notification.

We are using the "addHost", "editHost" and "delHost" methods to update
the hosttree, but it can be changed later to be more specific (addInterface,
addService, and so on)