Moving the creation of objetos to controller
The controller was registering the classes to the factory, but never
using it.
The api should be only a bridge between the outside world and our app,
so the api now uses the methods inside the controller.
Also, it was created a client and the routes to creates objects through
the REST api