Fix #657
The problem was that we are not adding the hosts to the workspace, and
when we were calling model_controller.getAllHosts we were using the
workspace's getHosts method.

Now we added a new method in the mappers to retrieve all the objetcs
from that mapper, and we are using it in the controller.