Fix incorrect already existing message in create_vuln fplugin
When I created the same vulnerability to a different host it incorrectly
said that the vulnerability already existed. I refactored the code so
this doesn't happen anymore. In addition, better error messages are
shown.

I also removed the `--dry-run` option, since it doesn't make any sense
now because we can't calculate the object ID without creating it first.