Remove ~/.faraday/plugins logic
The usage of this directory made no sense. It was designed to allow
users to write and modify plugins. But each time a user ran
faraday-client without the `--dev-mode` option, this directory was
silently removed.

Removing everything related to this directory seemed to be the best
choice. Also, it will prevent bugs related to filesystem permissions
when the source directory is read-only.